requirements.txt
Django==1.8.5
Jinja2==2.8
MarkupSafe==0.23
PyYAML==3.11
Twisted==15.4.0
argh==0.26.1
argparse==1.2.1
beautifulsoup4==4.4.1
cffi==1.3.0
characteristic==14.3.0
cryptography==1.1
django-cors-headers==1.1.0
django-geojson==2.8.1
django-leaflet==0.16.0
djangorestframework==3.1.1
enum34==1.0.4
hendrix==2.0.0
idna==2.0
ipaddress==1.0.14
jsonfield==1.0.3
mechanize==0.2.5
pathtools==0.1.2
project-runpy==0.3.1
psycopg2==2.5.1
pyOpenSSL==0.15.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pychalk==0.0.5
pycparser==2.14
requests==2.8.1
service-identity==14.0.0
six==1.10.0
txsockjs==1.2.2
watchdog==0.8.3
wsgiref==0.1.2
zope.interface==4.1.3
